requirements.txt
flask
kubernetes
requests
flask-socketio 
eventlet
huggingface_hub
PyYAML
boto3
langchain
faiss-cpu
python-dotenv
pillow